Develop

+++ src/app/Http/Controllers/Api/BookController.php
@@ -74,7 +74,7 @@ class BookController extends Controller
             $this->authorize('isBookManager', $workspace);
 
             DB::transaction(function () use ($workspaceId, $request): void {
-                $bookCategory = BookCategory::where('name', $request->get('bookCategoryName'))->firstOrFail();
+                $bookCategory = BookCategory::where('name', $request->get('category'))->firstOrFail();
                 $book = Book::create([
                     'workspace_id' => $workspaceId,
                     'book_category_id' => $bookCategory->id,
@@ -104,6 +104,10 @@ class BookController extends Controller
             $this->authorize('isBookManager', $workspace);
             $book = Book::find($request->get('id'));
             $bookCategory = BookCategory::where('name', $request->get('category'))->first();
+
+            if (!$book) {
+                abort(401);
+            }
 
             if ($book->image_path) {
                 \Storage::delete($book->image_path);
@@ -153,7 +157,7 @@ class BookController extends Controller
             $this->authorize('affiliation', $workspace);
             $this->authorize('isBookManager', $workspace);
             DB::transaction(function () use ($request): void {
-                $request->collect('book_ids')->each(function ($bookId): void {
+                $request->collect('bookIds')->each(function ($bookId): void {
                     BookPurchaseApply::where('book_id', $bookId)->delete();
                     BookRentalApply::where('book_id', $bookId)->delete();
                     BookReview::where('book_id', $bookId)->delete();
